+++ examples/basic/elementaryBookkeepingEx1.hs
@@ -237,30 +237,30 @@ netIncomeTransfer'
 main :: IO ()
 main = do
     print exp3
-    writeBS "examples/result/csv/exp3.csv" exp3
+    writeBS "examples/basic/result/csv/exp3.csv" exp3
     print "---"
     print q3A
-    writeBS "examples/result/csv/q3.csv" q3A
+    writeBS "examples/basic/result/csv/q3.csv" q3A
     print "---"
     print exp4
-    writePL "examples/result/csv/exp4.csv" exp4
+    writePL "examples/basic/result/csv/exp4.csv" exp4
     print "---"
     print q4A
     print q4A'
-    writeBS "examples/result/csv/q4.csv" q4A
+    writeBS "examples/basic/result/csv/q4.csv" q4A
     print "---"
     print exp5_A
-    writeJournal "examples/result/csv/exp5.csv" exp5_A getDay
+    writeJournal "examples/basic/result/csv/exp5.csv" exp5_A getDay
     print "---"
     print cashDebitTotal
     print cashCreditTotal
     print cashBalance1
     print cashBalance2
-    writeCompoundTrialBalance "examples/result/csv/q7.csv" q7
+    writeCompoundTrialBalance "examples/basic/result/csv/q7.csv" q7
     print "---"
     print exp8
     print $ incomeSummaryAccount exp8
     print $ netIncomeTransfer $ incomeSummaryAccount exp8
     print $ netIncomeTransfer' $ incomeSummaryAccount exp8
-    writePL "examples/result/csv/exp8A.csv" $ incomeSummaryAccount exp8
+    writePL "examples/basic/result/csv/exp8A.csv" $ incomeSummaryAccount exp8